Use the smaller num-traits dependency
There's no reason to pull in the larger `num` meta-crate, when only a
couple traits are needed.  On its own, `num-traits` now even supports
`#![no_std]` builds when the default `std` feature isn't enabled.

Note, this is not a breaking change, even though the traits are in the
public API of `float-cmp`, because they are the exact same that `num`
just re-exports.